commit4(final)
i was a bit confused with this one in the beginning bc of the warnings i was getting. i saw that i added unnecesary arguments in some of the functions before but i think i started to understand why and what happend.

i modified the load_unload_passengers function a bit. just added a max of 10 passengers and for the number of passengers to not go on negative.
also added smth in the main. if the user unloads all the passengers then the program ends.