(extra commit)
i though about how to make the elevator start from we left off even if we end the program. since we still havent worked with files yet, i checked it on the internet and it basiclly gave me all way how to write it. i thought it was cool so im including this here :)